+++ projects/ngx-scrollbar/src/lib/thumb/thumb.ts
@@ -7,18 +7,6 @@ import { ThumbAdapter } from './thumb-adapter';
   providers: [{ provide: ThumbAdapter, useExisting: ThumbXDirective }]
 })
 export class ThumbXDirective extends ThumbAdapter {
-
-  get offset(): number {
-    return this.clientRect.left;
-  }
-
-  get size(): number {
-    return this.nativeElement.offsetWidth;
-  }
-
-  protected get dragStartOffset(): number {
-    return this.offset + this.document.defaultView.scrollX;
-  }
 }
 
 @Directive({
@@ -27,16 +15,4 @@ export class ThumbXDirective extends ThumbAdapter {
   providers: [{ provide: ThumbAdapter, useExisting: ThumbYDirective }]
 })
 export class ThumbYDirective extends ThumbAdapter {
-
-  get offset(): number {
-    return this.clientRect.top;
-  }
-
-  get size(): number {
-    return this.nativeElement.offsetHeight;
-  }
-
-  protected get dragStartOffset(): number {
-    return this.offset + this.document.defaultView.scrollY;
-  }
 }