+++ projects/ngx-scrollbar/src/lib/track/track.ts
@@ -1,4 +1,4 @@
-import { Directive } from '@angular/core';
+import { Directive, effect } from '@angular/core';
 import { Observable } from 'rxjs';
 import { fromPromise } from 'rxjs/internal/observable/innerFrom';
 import { TrackAdapter } from './track-adapter';
@@ -12,81 +12,46 @@ export class TrackXDirective extends TrackAdapter {
 
   protected readonly cssLengthProperty: string = '--track-x-length';
 
-  get offset(): number {
-    return this.clientRect.left;
-  }
-
-  get size(): number {
-    return this.nativeElement.offsetWidth;
-  }
-
   protected get viewportScrollSize(): number {
     return this.cmp.viewport.contentWidth;
   }
 
-  protected get viewportSize(): number {
-    return this.cmp.viewport.offsetWidth;
+  getCurrPosition: () => number;
+
+  getScrollDirection: (position: number) => 'forward' | 'backward';
+
+  constructor() {
+    effect(() => {
+      if (this.cmp.direction() === 'rtl') {
+        this.getCurrPosition = (): number => {
+          const offset: number = this.viewportScrollSize - this.viewportSize - this.control.viewportScrollOffset;
+          return offset * this.size / this.viewportScrollSize;
+        };
+        this.getScrollDirection = (position: number): 'forward' | 'backward' => {
+          return position < this.getCurrPosition() ? 'forward' : 'backward';
+        };
+      } else {
+        this.getCurrPosition = (): number => {
+          return this.control.viewportScrollOffset * this.size / this.viewportScrollSize
+        };
+        this.getScrollDirection = (position: number): 'forward' | 'backward' => {
+          return position > this.getCurrPosition() ? 'forward' : 'backward';
+        };
+      }
+    });
+    super();
   }
 
-  protected getScrollDirection(position: number): 'forward' | 'backward' {
-    if (this.cmp.direction() === 'rtl') {
-      return position - this.thumb.offset < 0 ? 'forward' : 'backward';
-    }
-    return position - this.thumb.offset > 0 ? 'forward' : 'backward';
+  protected scrollTo(value: number): Observable<void> {
+    return fromPromise(this.cmp.scrollTo({ start: value, duration: this.cmp.trackScrollDuration }));
   }
 
-  protected scrollTo(left: number): Observable<void> {
-    return fromPromise(this.cmp.scrollTo({ left, duration: this.cmp.trackScrollDuration }));
-  }
-
-  protected getScrollForwardIncrement(): number {
-    if (this.cmp.direction() === 'rtl') {
-      const position: number = -(this.control.viewportScrollOffset - this.viewportSize);
-      return this.scrollMax - position;
-    }
+  protected getScrollForwardStep(): number {
     return this.control.viewportScrollOffset + this.viewportSize;
   }
 
-  protected getScrollBackwardIncrement(): number {
-    if (this.cmp.direction() === 'rtl') {
-      const position: number = -(this.control.viewportScrollOffset + this.viewportSize);
-      return this.scrollMax - position;
-    }
+  protected getScrollBackwardStep(): number {
     return this.control.viewportScrollOffset - this.viewportSize;
-  }
-
-  protected getOnGoingScrollForward(): { position: number, nextPosition: number, endPosition: number } {
-    if (this.cmp.direction() === 'rtl') {
-      const position: number = this.control.viewportScrollOffset - this.viewportSize;
-      return {
-        position,
-        nextPosition: this.scrollMax + position,
-        endPosition: 0
-      }
-    }
-    const position: number = this.control.viewportScrollOffset + this.viewportSize;
-    return {
-      position,
-      nextPosition: position,
-      endPosition: this.scrollMax
-    }
-  }
-
-  protected getOnGoingScrollBackward(): { position: number, nextPosition: number, endPosition: number } {
-    if (this.cmp.direction() === 'rtl') {
-      const position: number = this.control.viewportScrollOffset + this.viewportSize;
-      return {
-        position,
-        nextPosition: this.scrollMax + position,
-        endPosition: this.scrollMax
-      }
-    }
-    const position: number = this.control.viewportScrollOffset - this.viewportSize;
-    return {
-      position,
-      nextPosition: position,
-      endPosition: 0
-    }
   }
 }
 
@@ -99,53 +64,27 @@ export class TrackYDirective extends TrackAdapter {
 
   protected readonly cssLengthProperty: string = '--track-y-length';
 
-  get offset(): number {
-    return this.clientRect.top;
-  }
-
-  get size(): number {
-    return this.nativeElement.offsetHeight;
-  }
-
   protected get viewportScrollSize(): number {
     return this.cmp.viewport.contentHeight;
   }
 
-  protected get viewportSize(): number {
-    return this.cmp.viewport.offsetHeight;
+  protected getCurrPosition(): number {
+    return this.control.viewportScrollOffset * this.size / this.viewportScrollSize;
   }
 
   protected getScrollDirection(position: number): 'forward' | 'backward' {
-    return position - this.thumb.offset > 0 ? 'forward' : 'backward';
+    return position > this.getCurrPosition() ? 'forward' : 'backward';
   }
 
   protected scrollTo(top: number): Observable<void> {
     return fromPromise(this.cmp.scrollTo({ top, duration: this.cmp.trackScrollDuration }));
   }
 
-  protected getScrollForwardIncrement(): number {
+  protected getScrollForwardStep(): number {
     return this.control.viewportScrollOffset + this.viewportSize;
   }
 
-  protected getScrollBackwardIncrement(): number {
+  protected getScrollBackwardStep(): number {
     return this.control.viewportScrollOffset - this.viewportSize;
-  }
-
-  protected getOnGoingScrollForward(): { position: number, nextPosition: number, endPosition: number } {
-    const position: number = this.control.viewportScrollOffset + this.viewportSize;
-    return {
-      position,
-      nextPosition: position,
-      endPosition: this.scrollMax
-    }
-  }
-
-  protected getOnGoingScrollBackward(): { position: number, nextPosition: number, endPosition: number } {
-    const position: number = this.control.viewportScrollOffset - this.viewportSize;
-    return {
-      position,
-      nextPosition: position,
-      endPosition: 0
-    }
   }
 }